Refactor hcs internal pkg k/k repo using ContainerProperties and NetworkStats struct from hcs internal pkg. Due to this the whole hcsshim pkg is having to be imported in k/k. This commit helps to restructure the pkg and maintain submodule to ensure that only this pkg is imported in k/k
