+++ hcs/go.mod
@@ -1,0 +1,7 @@
+module github.com/Microsoft/hcsshim/hcs
+
+go 1.21
+
+require github.com/Microsoft/go-winio v0.6.2
+
+require golang.org/x/sys v0.10.0 // indirect
