+++ hcs/helpers.go
@@ -1,0 +1,7 @@
+package hcs
+
+import (
+	hcsschema "github.com/Microsoft/hcsshim/hcs/internal/schema2"
+)
+
+type Version = hcsschema.Version
