+++ internal/hcs/schema2/virtual_machine.go
@@ -11,9 +11,13 @@
 
 package hcsschema
 
+import(
+	hcstypes "github.com/Microsoft/hcsshim/hcs"
+)
+
 // Configuration of a virtual machine, used during its creation to set up and/or use resources.
 type VirtualMachine struct {
-	Version *Version `json:"Version,omitempty"`
+	Version *hcstypes.Version `json:"Version,omitempty"`
 	// When set to true, the virtual machine will treat a reset as a stop, releasing resources and cleaning up state.
 	StopOnReset     bool             `json:"StopOnReset,omitempty"`
 	Chipset         *Chipset         `json:"Chipset,omitempty"`
