+++ internal/hcs/system.go
@@ -12,6 +12,7 @@ import (
 	"syscall"
 	"time"
 
+	hcstypes "github.com/Microsoft/hcsshim/hcs"
 	"github.com/Microsoft/hcsshim/internal/cow"
 	"github.com/Microsoft/hcsshim/internal/hcs/schema1"
 	hcsschema "github.com/Microsoft/hcsshim/internal/hcs/schema2"
@@ -165,7 +166,7 @@ func (computeSystem *System) IsOCI() bool {
 }
 
 // GetComputeSystems gets a list of the compute systems on the system that match the query
-func GetComputeSystems(ctx context.Context, q schema1.ComputeSystemQuery) ([]schema1.ContainerProperties, error) {
+func GetComputeSystems(ctx context.Context, q schema1.ComputeSystemQuery) ([]hcstypes.ContainerProperties, error) {
 	operation := "hcs::GetComputeSystems"
 
 	queryb, err := json.Marshal(q)
@@ -182,7 +183,7 @@ func GetComputeSystems(ctx context.Context, q schema1.ComputeSystemQuery) ([]sch
 	if computeSystemsJSON == "" {
 		return nil, ErrUnexpectedValue
 	}
-	computeSystems := []schema1.ContainerProperties{}
+	computeSystems := []hcstypes.ContainerProperties{}
 	if err = json.Unmarshal([]byte(computeSystemsJSON), &computeSystems); err != nil {
 		return nil, err
 	}
@@ -344,7 +345,7 @@ func (computeSystem *System) ExitError() error {
 }
 
 // Properties returns the requested container properties targeting a V1 schema container.
-func (computeSystem *System) Properties(ctx context.Context, types ...schema1.PropertyType) (*schema1.ContainerProperties, error) {
+func (computeSystem *System) Properties(ctx context.Context, types ...schema1.PropertyType) (*hcstypes.ContainerProperties, error) {
 	computeSystem.handleLock.RLock()
 	defer computeSystem.handleLock.RUnlock()
 
@@ -368,7 +369,7 @@ func (computeSystem *System) Properties(ctx context.Context, types ...schema1.Pr
 	if propertiesJSON == "" {
 		return nil, ErrUnexpectedValue
 	}
-	properties := &schema1.ContainerProperties{}
+	properties := &hcstypes.ContainerProperties{}
 	if err := json.Unmarshal([]byte(propertiesJSON), properties); err != nil {
 		return nil, makeSystemError(computeSystem, operation, err, nil)
 	}
