+++ internal/schemaversion/schemaversion.go
@@ -7,28 +7,28 @@ import (
 	"encoding/json"
 	"fmt"
 
-	hcsschema "github.com/Microsoft/hcsshim/internal/hcs/schema2"
+	hcstypes "github.com/Microsoft/hcsshim/hcs"
 	"github.com/Microsoft/hcsshim/osversion"
 	"github.com/sirupsen/logrus"
 )
 
 // SchemaV10 makes it easy for callers to get a v1.0 schema version object
-func SchemaV10() *hcsschema.Version {
-	return &hcsschema.Version{Major: 1, Minor: 0}
+func SchemaV10() *hcstypes.Version {
+	return &hcstypes.Version{Major: 1, Minor: 0}
 }
 
 // SchemaV21 makes it easy for callers to get a v2.1 schema version object
-func SchemaV21() *hcsschema.Version {
-	return &hcsschema.Version{Major: 2, Minor: 1}
+func SchemaV21() *hcstypes.Version {
+	return &hcstypes.Version{Major: 2, Minor: 1}
 }
 
 // SchemaV25  makes it easy for callers to get a v2.5 schema version object.
-func SchemaV25() *hcsschema.Version {
-	return &hcsschema.Version{Major: 2, Minor: 5}
+func SchemaV25() *hcstypes.Version {
+	return &hcstypes.Version{Major: 2, Minor: 5}
 }
 
 // isSupported determines if a given schema version is supported
-func IsSupported(sv *hcsschema.Version) error {
+func IsSupported(sv *hcstypes.Version) error {
 	if IsV10(sv) {
 		return nil
 	}
@@ -50,7 +50,7 @@ func IsSupported(sv *hcsschema.Version) error {
 
 // IsV10 determines if a given schema version object is 1.0. This was the only thing
 // supported in RS1..3. It lives on in RS5, but will be deprecated in a future release.
-func IsV10(sv *hcsschema.Version) bool {
+func IsV10(sv *hcstypes.Version) bool {
 	if sv.Major == 1 && sv.Minor == 0 {
 		return true
 	}
@@ -60,7 +60,7 @@ func IsV10(sv *hcsschema.Version) bool {
 // IsV21 determines if a given schema version object is 2.0. This was introduced in
 // RS4, but not fully implemented. Recommended for applications using HCS in RS5
 // onwards.
-func IsV21(sv *hcsschema.Version) bool {
+func IsV21(sv *hcstypes.Version) bool {
 	if sv.Major == 2 && sv.Minor == 1 {
 		return true
 	}
@@ -68,7 +68,7 @@ func IsV21(sv *hcsschema.Version) bool {
 }
 
 // V25 schema introduced much later. Required to support SNP.
-func IsV25(sv *hcsschema.Version) bool {
+func IsV25(sv *hcstypes.Version) bool {
 	if sv.Major == 2 && sv.Minor == 5 {
 		return true
 	}
@@ -76,7 +76,7 @@ func IsV25(sv *hcsschema.Version) bool {
 }
 
 // String returns a JSON encoding of a schema version object
-func String(sv *hcsschema.Version) string {
+func String(sv *hcstypes.Version) string {
 	b, err := json.Marshal(sv)
 	if err != nil {
 		return ""
@@ -86,7 +86,7 @@ func String(sv *hcsschema.Version) string {
 
 // DetermineSchemaVersion works out what schema version to use based on build and
 // requested option.
-func DetermineSchemaVersion(requestedSV *hcsschema.Version) *hcsschema.Version {
+func DetermineSchemaVersion(requestedSV *hcstypes.Version) *hcstypes.Version {
 	sv := SchemaV10()
 	if osversion.Build() >= osversion.RS5 {
 		sv = SchemaV21()
